fix: change getWorkspaceId return type to uint64_t to prevent overflow
The monitorId is uint64_t but the return was truncated to uint32_t,
which could cause workspace ID collisions for large monitor IDs.

+++ include/TagsMonitor.hpp
@@ -43,7 +43,7 @@ class TagsMonitor {
      */
     void        unregisterCurrentWindow();
 
-    uint32_t    getWorkspaceId(uint16_t tag) const;
+    uint64_t    getWorkspaceId(uint16_t tag) const;
 
     static bool isValidTag(uint16_t tag);
